requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
autopep8==1.5.6
backcall==0.2.0
base58==2.1.0
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.6
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
multidict==5.1.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.8.1
PyQRCode==1.2.1
python3-indy==1.16.0
requests==2.25.1
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
yarl==1.6.3
